requirements.txt
# Core dependencies for Orchestrator Agent
google-generativeai>=0.8.0    # Google Gemini AI models
google-adk>=0.1.0             # Google Agent Development Kit
requests>=2.32.0              # HTTP requests for agent communication
uvicorn>=0.30.0               # ASGI server for running the agent
fastapi>=0.115.0              # Web framework (if needed by google-adk)
litellm>=1.49.0               # LiteLLM for local model support
redis>=5.1.0                  # Redis for data storage
a2a-sdk>=0.1.0                # Agent-to-Agent communication

# Environment and configuration
python-dotenv>=1.0.0          # Environment variable management (.env files)
pyyaml>=6.0.0                 # YAML configuration support

# Development and testing
pytest>=8.3.0                 # Testing framework
pytest-asyncio>=0.24.0        # Async testing support

# Optional: Ollama client (for local LLM fallback)
# ollama>=0.3.0

# Additional utilities
httpx>=0.27.0                 # Async HTTP client (used by agents)